Fix build errors after rebase by moving Linux specific for IPMASQ from main.go to ipmasq.go
